Arrows now in boxes; example is longer; words so far carries over
Note: ibex has limited communication between elements, and only between neighboring elements, so for words so far to be cumulative, all elements in between maze items have to copy the "counter" value from utils.getValuesFromPrevious to utils.setValuesForNextElement. This has been done for Maze_Separator (otherwise identical to Separator).

words so far is all words that have been correct, it does not reset to 0 when a mistake is made